fix(mobile): harden iOS interaction diagnostics before device validation
These changes make the iOS interaction diagnostics page safer to run on a physical device after simulator validation. The keyboard viewport state is now shared from MobileLayout via an injection key, so the diagnostics page reuses the existing listener instead of creating a second one. Diagnostic cold-entry helpers are moved from the boot file into src/diagnostics/ so production boot code does not depend on diagnostics internals. Capacitor resume listener registration errors are now swallowed in the boot path so a native plugin failure cannot block launch. The IME probe blur trigger switches from click to pointerdown, and CLAUDE.md records the two behaviors still needing physical-device confirmation.
